fix vblank timings, fix apu counter ticking
previously, i was ticking each channel, regardless if the counter was already
zero.
this meant that when a game set the tone to 0 (but kept vol on), then the
polarity would rappidly change, causing a horrible screech.

as for the vdp, i didn't take into account that the visible screen is actually
somewhat centered, as in, it doesnt start at line 0, but theres actually a border.

finally, because vcount is 8-bit, that means it'll eventually overflow, as the
screen lines is 262 (for ntsc).
what the sms actually does is when it hits a certain line, it sets vcount to a
lower value, so that it never overflows.