fix use-after-free bug in web savestate, fix savestate with sram
there was a very bad bug in the web savestate code.

for the rom name, i was just saving a copy of the ptr to the path_str. this was
fine for desktop as i was using argv, however, for web, the name was allocated,
and then freed after the loadrom function!

this magically worked for a while, but after restarting the browser, it no longer
worked, of course, because that block of memory has since been overwritten so the
garbage name it did store (used for state path) is no longer valid!

the second bug was caused by loading state of sram games when they had sram currently
mapped, because on state load, i re-map all the rom banks, but i didn't check if sram
was actually mapped.