many changes, see description
sg1000 support! i think every game works fine (up until the title screen that is).
render support isn't finished yet either.

ported the code to a few consoles, such as n64 and gamecube. none of the ports are in a finished state,
most are just the boiler plate code and testing performance.

added an unfinished database for all sms, gg and sg releases. this is needed for knowing the mapper of the game
as well as sram support. another unused benefit is knowing the size of sram, though maybe this isn't important as
mekka has been fine all these by assuming the max sram size for all games.

almost all mappers are implemented, only mapper left is the codemasters.

i started on a debugger, basically taking the code from my gba emulator and hooking it to this.
currently, there's only a graph for psg playback, not usefull at all, kinda cool to look at though.

i pulled in a helper lib i wrote called mgb (my game boy, the name of my first emu), which provides common
functions such as rom loading, savestates, rewind, keeping track of if a rom is loaded.
basically acts like a stdlib for emulators.
i do plan to make that lib generic, so that the same code can be used across all my emulators, with a single
file needed to be written per one which adds the names of functions to call, filters for rom extensions etc.

i corrected some of timings in my z80.c.

fixes #26, #23, #14