fix single_file building, optimise audio sample callback, remove "debugger"
removed the debugger because it was not finished and pretty useless.

optimised the audio callback as now you can pass in a buffer which the core will first fill
and then call the callback. this reduces function call overhead.

added basic mixer function to the core for s16 output.
tis function will be improved for accuracy eventually with proper mixing as
it is unlikely that each channel is perfectly mixed (see NES APU).