add sram dirty flag to reduce the number of sram writes to disk.
this isn't perfect as writes to sram aren't tracked, so the dirty flag will always be set so long
as sram is mapped.
this is done to remove overhead on writes as otherwise all writes would have the cost of checking if its an sram write.

a better solution is for mgb (frontend) to use mmap for sram, however that would require changes to the core.