initial pal support.
pal seems to work, tested with New Zeland Story and Sonic (sms) and they both play at the correct rate.
support for pal only games needs to be added to the rom database.

pal still needs to be added to the cli args list, as well as more testing with code master games.